+++ src/main/java/com/github/toolarium/dependency/check/DependencyCheckUtil.java
@@ -37,7 +37,7 @@ public final class DependencyCheckUtil {
      *
      * @author patrick
      */
-    private static class HOLDER {
+    private static final class HOLDER {
         static final DependencyCheckUtil INSTANCE = new DependencyCheckUtil();
     }