+++ src/main/java/com/github/toolarium/dependency/check/formatter/DependencyCheckFormatterFactory.java
@@ -20,7 +20,7 @@ public final class DependencyCheckFormatterFactory {
      *
      * @author patrick
      */
-    private static class HOLDER {
+    private static final class HOLDER {
         static final DependencyCheckFormatterFactory INSTANCE = new DependencyCheckFormatterFactory();
     }